+++ lib/provider.dart
@@ -4,10 +4,12 @@ import 'controllers/check_decibel_page.controller.dart';
 import 'controllers/game_screen_controller.dart';
 import 'controllers/instrument_page_controller.dart';
 import 'controllers/login_page_controller.dart';
+import 'controllers/music_controller.dart';
 
 void initControllers() {
   Get.put(LoginPageController());
   Get.put(InstrumentPageController());
   Get.put(CheckDecibelPageController());
   Get.put(GameScreenController());
+  Get.put(MusicController());
 }
